Decompose ConnectionSettingsView into one section per transport.
The 1544-line god component held the UI for all four transports (serial,
socket, RTT, Bluetooth LE) plus the shared chrome. Split each transport
into its own file so future UX work on one transport doesn't have to
scroll past the others, and so adding a fifth (or replacing one) is a
contained change:

- SerialPortSettingsSection.tsx (366 LOC)
- SocketSettingsSection.tsx     (208 LOC)
- RttSettingsSection.tsx        (330 LOC)
- BluetoothLESettingsSection.tsx (491 LOC)

ConnectionSettingsView.tsx is now a 177-line router: the connection-type
select + a switch on connType + the General Settings footer (which has
per-transport tweaks to its label/tooltip wording, so it stays at the
parent level).

Per-section state moves to its owning component:
- the local socket-port input lives in SocketSettingsSection
- the RTT log autoscroll ref + RTT path auto-fill effect live in
  RttSettingsSection (only run while RTT is selected, matching the
  pre-refactor behaviour)
- the BLE-only `getDisplayUuids` and `formatManufacturerData` helpers
  move to BluetoothLESettingsSection
- the parent retains only `scanForSerialPorts()` on mount, because it
  runs unconditionally regardless of selected transport.

Behaviour is unchanged — pure refactor. 140/140 unit tests still pass,
tsc clean, no new lint errors (warning count dropped 307→303 from
incidental cleanup of unused imports).

No CHANGELOG entry per the carve-out for internal-only refactors with
zero outward impact.
